Added basic group chat functionality, including group roles and basic admin stuff

+++ app/controllers/group_controller.rb
@@ -5,8 +5,9 @@ class GroupController < ApplicationController
       user=current_user
       group = Group.create(groupname: params[:group_name])
       user.groups << group
+      user.set_role("Admin", group)
       flash[:message]="#{group.groupname} was created successfully"
-      redirect "/users"
+      redirect "/groups/#{group.id}"
     else
       flash[:message]="Not logged in"
       redirect '/'
@@ -17,9 +18,20 @@ class GroupController < ApplicationController
     if logged_in?
       user=current_user
       group = Group.find(params[:id].to_i)
-      user.groups << group
-      flash[:message] = "Successfully joined #{group.groupname}"
-      redirect "/groups/#{group.id}"
+      role=user.get_role(group)
+      if role.nil?
+        user.groups << group
+        user.set_role("Member", group)
+        group.messages.create(themessage: "#{user.username} has joined the group", user: current_user)
+        flash[:message] = "Successfully joined #{group.groupname}"
+        redirect "/groups/#{group.id}"
+      elsif role=="Kicked"
+        flash[:message] = "You were kicked out and cannot rejoin this group."
+        redirect "/groups/#{group.id}"
+      elsif role == "Member"
+        flash[:message] = "You are already a member of this group"
+        redirect "/groups/#{group.id}"
+      end
     else
       flash[:message]="Not logged in"
       redirect '/'
@@ -29,8 +41,9 @@ class GroupController < ApplicationController
   get '/groups/:id' do
     if logged_in?
       @user = current_user
-      @groups = @user.groups_without
+      @not_in_groups = @user.not_in_groups
       @group = Group.find(params[:id].to_i)
+      @role = @user.get_role(@group)
       erb :'/groups/show'
     else
       flash[:message]="Not logged in"
@@ -40,19 +53,50 @@ class GroupController < ApplicationController
 
   delete '/groups/:id/delete' do
     group = Group.find(params[:id].to_i)
-    name=group.groupname
-    group.destroy
-    flash[:message]="#{name} was deleted!"
-    redirect '/users'
+    role=current_user.get_role(group)
+    if role != "Admin"
+      flash[:message]="You are not the administrator of this group"
+      redirect "/groups/#{group.id}"
+    else
+      name=group.groupname
+      group.destroy
+      flash[:message]="#{name} was deleted!"
+      redirect '/users'
+    end
+  end
+
+  get '/groups/:groupid/user/:userid/kick' do
+    user=User.find(params[:userid].to_i)
+    group=Group.find(params[:groupid].to_i)
+    user.set_role("Kicked", group)
+    group.messages.create(themessage: "#{user.username} was kicked out of the group", user: current_user)
+    flash[:message] = "#{user.username} was successfully kicked from #{group.groupname}"
+    redirect "/groups/#{group.id}"
+  end
+
+  get '/groups/:groupid/user/:userid/reinstate' do
+    user=User.find(params[:userid].to_i)
+    group=Group.find(params[:groupid].to_i)
+    user.set_role("Member", group)
+    group.messages.create(themessage: "#{user.username} was invited back", user: current_user)
+    flash[:message] = "#{user.username} was invited back to #{group.groupname}"
+    redirect "/groups/#{group.id}"
   end
 
   patch '/groups/:id/leave' do
     if logged_in?
       user = current_user
       group = Group.find(params[:id].to_i)
-      user.groups.delete(group)
-      flash[:message]="Successfully left #{group.groupname}"
-      redirect "/groups/#{group.id}"
+      role = user.get_role(group)
+      if role == "Admin"
+        flash[:message]="You can't leave.  You're the Admin."
+        redirect "/groups/#{group.id}"
+      else
+        user.groups.delete(group)
+        group.messages.create(themessage: "#{user.username} has left the group", user: current_user)
+        flash[:message]="Successfully left #{group.groupname}"
+        redirect "/groups/#{group.id}"
+      end
     else
       flash[:message]="Not logged in"
       redirect '/'
